Implement screen orientation changes (broken)
Introduced the capability to adjust screen orientation by adding a function to set the screen size based on the new orientation and modifying the setup of the Panel_PaxoSDL. Also updated references to lgfx::Panel_sdl to Panel_PaxoSDL to maintain consistency. Future work will be needed to complete the rotation functionality and iron out details.