Update window size setting in Panel_PaxoSDL (still broken)
The width and height of the window are now set directly in the Panel_PaxoSDL class. Changes include modification to the setWindowSize function to use local variables rather than the configuration settings. Also a call to Panel_PaxoSDL::loop() was added in SDLInit to ensure the window is created before a new thread is started. Furthermore, print statement for debugging window size is included.