Add landscape buffer to handle color glitch
A "landscape buffer" was added to the graphics.cpp file in order to manage a color glitch that occurs when using the "setRotation()" function in LovyanGFX. This buffer serves as a temporary solution to this issue and prevents the glitch from appearing. The code has been adapted to use the new landscape buffer and the rotation has also been updated accordingly.