Refactor Keyboard class and add keyboard test app
Makes several changes to the Keyboard class in gui/src/elements, such as the conditional setting of screen dimensions based on orientation and the deprecation of the SImage class in favor of the Image class. Additionally, introduces a new 'keyboard_test' application to test the updated functionality of the Keyboard class.