Add custom fontSize support and default text functionality
Updated the GUI and Keyboard classes to introduce custom fontSize support for drawing texts on the canvas. Made improvements to the Keyboard class by allowing default text to be set when initializing the keyboard. This update also includes resetting the keyboard text to the default when the exit button is touched.